Merge Billing report into Project Details report.
The Billing report was causing some concerns and had largely common
functionality with the Project Details report, so the simple solution
was to combine the two with no mention of "Billing" anywhere.

Also removed the pie chart from Project Details. The improvements to the
pie chart in utilisation.js in 83a382ab90ddf367a311689ab0504717c0dba0c2
made made this pie chart redundant. Only difference is that it is no
longer possible to see resource quotas aggregated over whole
institutions, but that is something that should be added to
utilisation.js if needed.

Add date selection fields and walltime column to Project Details.

Add usage note for historical usage chart for Utilisation.

Fix table to list volumes when appropriate.

Change-Id: Idbdd9f886b192cb6540aee67ee2eaf5e2ba73692