Fix date range bugs.
Switching project/organisation would not correctly update the value of
the date range selectors, because they were being "updated" too soon.

This could also cause the walltime values to be computed incorrectly.

Also made the behaviour of the end date picker less confusing, so the
date no longer jumps to the next day after being selected (but this
comes at the cost of one millisecond's accuracy)...

Change-Id: I563c804cc0cf587833a77c00ee3f99b05708ce7a