Use aggregate_host historical data
Look up hypervisors' availability_zone values using historical mappings
in the aggregate_host table. Without this change, if an instance is
created on a host in availability_zone AZ1, then deleted and later the
host is moved to AZ2, the instance would incorrectly be shown as having
existed in AZ2.

Change-Id: I4ca2fc24b80702fd0c967172f30679c27b63a1ae